Rename TableFooter module to reflect the components name
Use same name for module as its component.
