+++ src/web/pages/hosts/Identifiers.jsx
@@ -11,8 +11,8 @@ import DeleteIcon from 'web/components/icon/DeleteIcon';
 import DetailsLink from 'web/components/link/DetailsLink';
 import TableBody from 'web/components/table/Body';
 import TableData from 'web/components/table/Data';
-import TableFooter from 'web/components/table/Footer';
 import Table from 'web/components/table/StripedTable';
+import TableFooter from 'web/components/table/TableFooter';
 import TableHead from 'web/components/table/TableHead';
 import TableHeader from 'web/components/table/TableHeader';
 import TableRow from 'web/components/table/TableRow';
